+++ tissuumaps/static/js/flask.js
@@ -50,7 +50,7 @@ flask.standalone.init = function () {
     // as long as the user is interacting with the page
     flask.standalone.pixelFlickering = HTMLElementUtils.createElement({"kind":"div", extraAttributes:{"style":"position:absolute;right:0px;bottom:0px;width:1px;height:1px;line-height:1px;background-color:#FFFFFF"}});
     document.body.append(flask.standalone.pixelFlickering)
-    var counterFrames = 3;
+    var counterFrames = 10;
     async function refreshAnimation() {
         // console.log("refreshAnimation", counterFrames);
         if (flask.standalone.pixelFlickering.style.backgroundColor=="rgb(255, 255, 255)")
@@ -66,12 +66,13 @@ flask.standalone.init = function () {
     }
     refreshAnimation();
     function resetCounter() {
-        counterFrames = 3;
+        counterFrames = 10;
     }
     document.addEventListener('click', resetCounter);
     document.addEventListener('mouseup', resetCounter);
     document.addEventListener('mousedown', resetCounter);
     document.addEventListener('mousemove', resetCounter);
+    document.addEventListener('pointermove', resetCounter);
     document.addEventListener('touchstart', resetCounter);
     document.addEventListener('scroll', resetCounter);
     document.addEventListener('keydown', resetCounter);
@@ -93,6 +94,17 @@ flask.standalone.init = function () {
         return false;
     }, true);
 };
+
+flask.standalone.addGeoJSON = function (filename) {
+    const queryString = window.location.search;
+    const urlParams = new URLSearchParams(queryString);
+    const path = urlParams.get('path')
+    flask.standalone.backend.addGeoJSON(path, filename, function(geoJSONJSON) {
+        if (geoJSONJSON["geoJSONPath"]!=null) {
+            regionUtils.JSONToRegions(geoJSONJSON["geoJSONPath"]);
+        }
+    });
+}
 
 flask.standalone.addCSV = function (filename) {
     const queryString = window.location.search;