+++ tissuumaps/static/js/utils/glUtils.js
@@ -19,6 +19,7 @@ glUtils = {
     _vaos: {},
     _textures: {},
     _query: null,
+    _caps: {},
 
     // Marker settings and info stored per UID (this could perhaps be
     // better handled by having an object per UID that stores all info
@@ -64,8 +65,9 @@ glUtils = {
     _showEdgesExperimental: true,
     _edgeThicknessRatio: 0.1,     // Ratio between edge thickness and marker size
     _regionOpacity: 0.5,
-    _regionFillRule: "never",   // Possible values: "never" | "nonzero" | "oddeven"
-    _regionUsePivotSplit: false,   // Use split edge lists for faster region rendering and less risk of overflow
+    _regionStrokeWidth: 1.0,      // Base stroke width in pixels (larger values can give artifacts, so use with care!)
+    _regionFillRule: "never",     // Possible values: "never" | "nonzero" | "oddeven"
+    _regionUsePivotSplit: false,  // Use split edge lists for faster region rendering and less risk of overflow
     _regionUseColorByID: false,   // Map region object IDs to unique colors
     _regionDataSize: {},          // Size stored per region data texture and used for dynamic resizing
     _regionPicked: null,          // Key to regionUtils._regions dict, or null if no region is picked
@@ -189,6 +191,7 @@ glUtils._markersVS = `
         // Discard point here in vertex shader if marker is hidden
         v_color.a = v_color.a > 0.0 ? in_opacity * u_markerOpacity : 0.0;
         v_color.a *= alphaFactorSize * alphaFactorSize;
+        v_color.a = clamp(v_color.a, 0.0, 1.0);
         if (v_color.a == 0.0) DISCARD_VERTEX;
     }
 `;
@@ -584,6 +587,7 @@ glUtils._regionsFS = `
 
     uniform int u_numScanlines;
     uniform float u_regionOpacity;
+    uniform float u_regionStrokeWidth;
     uniform int u_regionFillRule;
     uniform int u_regionUsePivotSplit;
     uniform int u_regionUseColorByID;
@@ -620,8 +624,13 @@ glUtils._regionsFS = `
         int scanline = int(v_scanline);
 
         float pixelWidth = length(dFdx(p.xy));
-        float strokeWidth = pixelWidth *
+        float strokeWidthPixels = u_regionStrokeWidth *
             (u_regionFillRule == FILL_RULE_NEVER ? STROKE_WIDTH : STROKE_WIDTH_FILLED);
+        // For proper anti-aliasing, clamp stroke width to at least 1 pixel, and
+        // make thinner strokes fade by coverage
+        float strokeWidth = max(1.0, strokeWidthPixels) * pixelWidth;
+        float strokeFade = min(1.0, strokeWidthPixels);
+
         float minEdgeDist = 1e7;  // Distance to closest edge
 
         vec4 scanlineInfo = texelFetch(u_regionData, ivec2(scanline, 0), 0);
@@ -682,9 +691,10 @@ glUtils._regionsFS = `
                         objectColor.rgb = scanDir > 0.0 ? vec3(0.0, 1.0, 0.0) : vec3(1.0, 0.0, 1.0);
                     }
                 #endif  // SHOW_PIVOT_SPLIT_DEBUG
-                    float edgeOpacity = smoothstep(strokeWidth, strokeWidth - pixelWidth, minEdgeDist);
-                    float fillOpacity = float(isInside) * u_regionOpacity;
-                    objectColor.a *= clamp(edgeOpacity + fillOpacity, 0.0, 1.0);
+                    float minEdgeDistSigned = isInside ? minEdgeDist : -minEdgeDist;
+                    float strokeOpacity = smoothstep(strokeWidth, strokeWidth - pixelWidth, minEdgeDist) * strokeFade;
+                    float fillOpacity = smoothstep(-pixelWidth, pixelWidth, minEdgeDistSigned) * u_regionOpacity;
+                    objectColor.a *= clamp(strokeOpacity + fillOpacity, 0.0, 1.0);
 
                     color.a = objectColor.a + (1.0 - objectColor.a) * color.a;
                     color.rgb = mix(color.rgb, objectColor.rgb, objectColor.a);
@@ -889,21 +899,23 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
 
     // Additional info about the vertex format. Make sure to update also
     // NUM_BYTES_PER_MARKER and NUM_BYTES_PER_MARKER_SECONDARY when making
-    // changes to the format! Two buffers will be used for the vertex data
-    // to avoid the 1GB max size limit imposed by QtWebEngine/Chromium.
+    // changes to the format! Two buffers will be used for the vertex data to
+    // avoid the 1GB max size limit imposed by QtWebEngine/Chromium. Also
+    // remember to update glUtils._updateBindingOffsetsForVAO if changing the
+    // format.
     const NUM_BYTES_PER_MARKER = 16;
     const NUM_BYTES_PER_MARKER_SECONDARY = 16;
     const POINT_OFFSET = numPoints * 0,
           INDEX_OFFSET = numPoints * 0,
           SCALE_OFFSET = numPoints * 4,
-          SHAPE_OFFSET = numPoints * 8;
-          OPACITY_OFFSET = numPoints * 12;
+          SHAPE_OFFSET = numPoints * 8,
+          OPACITY_OFFSET = numPoints * 12,
           TRANSFORM_OFFSET = numPoints * 14;
     const POINT_LOCATION = 0,
           INDEX_LOCATION = 1,
           SCALE_LOCATION = 2,
           SHAPE_LOCATION = 3,
-          OPACITY_LOCATION = 4;
+          OPACITY_LOCATION = 4,
           TRANSFORM_LOCATION = 5;
 
     const lastInputs = glUtils._markerInputsCached[uid];
@@ -911,6 +923,27 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
         dataUtils.data[uid]["modified"] = false;
         forceUpdate = true;
         scalarRange = [1e9, -1e9];  // This range will be computed from the data
+
+        console.time("Generate index data");
+        const numIndices = numPoints * numSectors;
+        let indicesSorted = new Uint32Array(numIndices);
+        {
+            for (let index = 0; index < numIndices; ++index) {
+                indicesSorted[index] = index;
+            }
+            if (useSortByCol) {
+                const colData = markerData[sortByCol];
+                if (sortByDesc) {
+                    // Sort in descending order
+                    indicesSorted.sort((i, j) => Number(colData[Math.floor(j / numSectors)]) -
+                        Number(colData[Math.floor(i / numSectors)]));
+                } else {  // Sort in ascending order
+                    indicesSorted.sort((i, j) => Number(colData[Math.floor(i / numSectors)]) -
+                        Number(colData[Math.floor(j / numSectors)]));
+                }
+            }
+        }
+        console.timeEnd("Generate index data");
 
         // Extract and upload vertex data for markers. For datasets with tens of of
         // millions of points, the vertex data can be quite large, so we upload the
@@ -940,7 +973,7 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
                 bytedata_transform = new Uint16Array(chunkSize * numSectors * 1);
 
                 for (let i = 0; i < chunkSize; ++i) {
-                    const markerIndex = i + offset;
+                    const markerIndex = indicesSorted[i + offset];
                     const sectors = markerData[sectorsPropertyName][markerIndex].toString().split(";");
                     const piechartAngles = glUtils._createPiechartAngles(sectors);
                     const lutIndex = (keyName != null) ? barcodeToLUTIndex[markerData[keyName][markerIndex]] : 0;
@@ -967,7 +1000,7 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
                 }
             } else {
                 for (let i = 0; i < chunkSize; ++i) {
-                    const markerIndex = i + offset;
+                    const markerIndex = indicesSorted[i + offset];
                     const lutIndex = (keyName != null) ? barcodeToLUTIndex[markerData[keyName][markerIndex]] : 0;
                     const opacity = useOpacityFromMarker ? markerData[opacityPropertyName][markerIndex] : 1.0;
                     if (useCollectionItemFromMarker) collectionItemIndex = markerData[collectionItemPropertyName][markerIndex];
@@ -1010,9 +1043,6 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
             if (!(uid + "_markers_secondary" in glUtils._buffers))
                 glUtils._buffers[uid + "_markers_secondary"] = glUtils._createVertexBuffer(
                     gl, numPoints * numSectors * NUM_BYTES_PER_MARKER_SECONDARY);
-            if (!(uid + "_markers_indices" in glUtils._buffers))
-                glUtils._buffers[uid + "_markers_indices"] = glUtils._createIndexBuffer(
-                    gl, numPoints * numSectors * 4);  // Use 32-bits indices
             if (!(uid + "_markers" in glUtils._vaos))
                 glUtils._vaos[uid + "_markers"] = gl.createVertexArray();
             if (!(uid + "_markers_instanced" in glUtils._vaos))
@@ -1041,14 +1071,6 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
                         gl.bufferData(gl.ARRAY_BUFFER, newBufferSize, gl.STATIC_DRAW);
                     gl.bindBuffer(gl.ARRAY_BUFFER, null);
                 }
-                {
-                    gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, glUtils._buffers[uid + "_markers_indices"]);
-                    const newBufferSize = numPoints * numSectors * 4;  // Use 32-bits indices
-                    const oldBufferSize = gl.getBufferParameter(gl.ELEMENT_ARRAY_BUFFER, gl.BUFFER_SIZE);
-                    if (newBufferSize != oldBufferSize)
-                        gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, newBufferSize, gl.STATIC_DRAW);
-                    gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, null);
-                }
             }
             gl.bindBuffer(gl.ARRAY_BUFFER, glUtils._buffers[uid + "_markers"]);
             gl.bufferSubData(gl.ARRAY_BUFFER, (POINT_OFFSET + offset * 16) * numSectors, bytedata_point);
@@ -1061,30 +1083,6 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
             gl.bindBuffer(gl.ARRAY_BUFFER, null);
         }
         console.timeEnd("Generate vertex data");
-
-        console.time("Generate index data");
-        {
-            const numIndices = numPoints * numSectors;
-            const bytedata_indices = new Uint32Array(numIndices);
-            for (let index = 0; index < numIndices; ++index) {
-                bytedata_indices[index] = index;
-            }
-            
-            if (useSortByCol) {
-                const colData = markerData[sortByCol];
-                if (sortByDesc) {
-                    // Sort in descending order
-                    bytedata_indices.sort((i, j) => Number(colData[j]) - Number(colData[i]));
-                } else {  // Sort in ascending order
-                    bytedata_indices.sort((i, j) => Number(colData[i]) - Number(colData[j]));
-                }
-            }
-
-            gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, glUtils._buffers[uid + "_markers_indices"]);
-            gl.bufferSubData(gl.ELEMENT_ARRAY_BUFFER, 0, bytedata_indices);
-            gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, null);
-        }
-        console.timeEnd("Generate index data");
 
         // Set up VAO with vertex format for drawing
         gl.bindVertexArray(glUtils._vaos[uid + "_markers"]);
@@ -1102,7 +1100,6 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
         gl.vertexAttribPointer(OPACITY_LOCATION, 1, gl.UNSIGNED_SHORT, true, 0, OPACITY_OFFSET * numSectors);
         gl.enableVertexAttribArray(TRANSFORM_LOCATION);
         gl.vertexAttribPointer(TRANSFORM_LOCATION, 1, gl.UNSIGNED_SHORT, false, 0, TRANSFORM_OFFSET * numSectors);
-        gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, glUtils._buffers[uid + "_markers_indices"]);
         gl.bindVertexArray(null);
 
         // Set up 2nd VAO (for experimental instanced drawing)
@@ -1127,7 +1124,6 @@ glUtils.loadMarkers = function(uid, forceUpdate) {
         gl.enableVertexAttribArray(TRANSFORM_LOCATION);
         gl.vertexAttribPointer(TRANSFORM_LOCATION, 1, gl.UNSIGNED_SHORT, false, 0, TRANSFORM_OFFSET * numSectors);
         gl.vertexAttribDivisor(TRANSFORM_LOCATION, 1);
-        gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, glUtils._buffers[uid + "_markers_indices"]);
         gl.bindVertexArray(null);
     }
     glUtils._markerInputsCached[uid] = JSON.stringify(newInputs);
@@ -1207,7 +1203,6 @@ glUtils.deleteMarkers = function(uid) {
     // Clean up WebGL resources
     gl.deleteBuffer(glUtils._buffers[uid + "_markers"]);
     gl.deleteBuffer(glUtils._buffers[uid + "_markers_secondary"]);
-    gl.deleteBuffer(glUtils._buffers[uid + "_markers_indices"]);
     gl.deleteVertexArray(glUtils._vaos[uid + "_markers"]);
     gl.deleteVertexArray(glUtils._vaos[uid + "_markers_instanced"]);
     gl.deleteBuffer(glUtils._buffers[uid + "_edges"]);
@@ -1216,7 +1211,6 @@ glUtils.deleteMarkers = function(uid) {
     gl.deleteTexture(glUtils._textures[uid + "_colorscale"]);
     delete glUtils._buffers[uid + "_markers"];
     delete glUtils._buffers[uid + "_markers_secondary"];
-    delete glUtils._buffers[uid + "_markers_indices"];
     delete glUtils._vaos[uid + "_markers"];
     delete glUtils._vaos[uid + "_markers_instanced"];
     delete glUtils._buffers[uid + "_edges"];
@@ -1228,6 +1222,38 @@ glUtils.deleteMarkers = function(uid) {
 
     // Make sure piechart legend is deleted if it was used for this UID
     markerUtils.updatePiechartLegend();
+}
+
+
+glUtils._updateBindingOffsetsForCurrentVAO = function(gl, uid, offset, numPoints) {
+    // This function is used for updating the offsets of vertex bindings for
+    // instanced drawing. This is only necessary because drawArrayInstanced and
+    // drawElementsInstanced ignores any offset or first parameters we provide
+    // for instanced arrays (arrays with non-zero vertex attrib divisor).
+
+    // Additional info about the vertex format. Make sure to update these values
+    // if/when also changing the vertex format in glUtils.loadMarkers!
+    const POINT_OFFSET = numPoints * 0,
+          INDEX_OFFSET = numPoints * 0,
+          SCALE_OFFSET = numPoints * 4,
+          SHAPE_OFFSET = numPoints * 8,
+          OPACITY_OFFSET = numPoints * 12,
+          TRANSFORM_OFFSET = numPoints * 14;
+    const POINT_LOCATION = 0,
+          INDEX_LOCATION = 1,
+          SCALE_LOCATION = 2,
+          SHAPE_LOCATION = 3,
+          OPACITY_LOCATION = 4,
+          TRANSFORM_LOCATION = 5;
+
+    gl.bindBuffer(gl.ARRAY_BUFFER, glUtils._buffers[uid + "_markers"]);
+    gl.vertexAttribPointer(POINT_LOCATION, 4, gl.FLOAT, false, 0, POINT_OFFSET + offset * 16);
+    gl.bindBuffer(gl.ARRAY_BUFFER, glUtils._buffers[uid + "_markers_secondary"]);
+    gl.vertexAttribIPointer(INDEX_LOCATION, 1, gl.INT, 0, INDEX_OFFSET + offset * 4);
+    gl.vertexAttribPointer(SCALE_LOCATION, 1, gl.FLOAT, false, 0, SCALE_OFFSET + offset * 4);
+    gl.vertexAttribPointer(SHAPE_LOCATION, 1, gl.FLOAT, false, 0, SHAPE_OFFSET + offset * 4);
+    gl.vertexAttribPointer(OPACITY_LOCATION, 1, gl.UNSIGNED_SHORT, true, 0, OPACITY_OFFSET + offset * 2);
+    gl.vertexAttribPointer(TRANSFORM_LOCATION, 1, gl.UNSIGNED_SHORT, false, 0, TRANSFORM_OFFSET + offset * 2);
 }
 
 
@@ -1934,13 +1960,13 @@ glUtils._drawMarkersByUID = function(gl, viewportTransform, markerScaleAdjusted,
     const numPoints = glUtils._numPoints[uid];
     if (numPoints == 0) return;
 
-    // Note: marker rendering is currently broken when both instancing and
-    // sorting are enabled at the same time. In that case, as workaround,
-    // we will fall back to using point sprites instead.
-    const useInstancing = glUtils._useInstancing && !glUtils._useSortByCol[uid];
+    // Chunk size used to split a single large draw call into smaller chunks. On
+    // some Android phones, drawing larger datasets can result in WebGL context
+    // loss or crashes, so this should be a workaround.
+    const chunkSize = 65536;
 
     // Set up render pipeline
-    const program = glUtils._programs[useInstancing ? "markers_instanced" : "markers"];
+    const program = glUtils._programs[glUtils._useInstancing ? "markers_instanced" : "markers"];
     gl.useProgram(program);
     gl.enable(gl.BLEND);
     gl.blendFuncSeparate(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA, gl.ONE, gl.ONE_MINUS_SRC_ALPHA);
@@ -1949,14 +1975,15 @@ glUtils._drawMarkersByUID = function(gl, viewportTransform, markerScaleAdjusted,
     gl.uniformMatrix2fv(gl.getUniformLocation(program, "u_viewportTransform"), false, viewportTransform);
     gl.uniform2fv(gl.getUniformLocation(program, "u_canvasSize"), [gl.canvas.width, gl.canvas.height]);
     gl.uniform1f(gl.getUniformLocation(program, "u_markerScale"), markerScaleAdjusted);
-    gl.uniform1f(gl.getUniformLocation(program, "u_maxPointSize"), useInstancing ? 2048 : 256);
+    gl.uniform1f(gl.getUniformLocation(program, "u_maxPointSize"),
+        glUtils._useInstancing ? 2048 : glUtils._caps[gl.ALIASED_POINT_SIZE_RANGE][1]);
     gl.uniformBlockBinding(program, gl.getUniformBlockIndex(program, "TransformUniforms"), 0);
     gl.bindBufferBase(gl.UNIFORM_BUFFER, 0, glUtils._buffers["transformUBO"]);
     gl.activeTexture(gl.TEXTURE2);
     gl.bindTexture(gl.TEXTURE_2D, glUtils._textures["shapeAtlas"]);
     gl.uniform1i(gl.getUniformLocation(program, "u_shapeAtlas"), 2);
 
-    gl.bindVertexArray(glUtils._vaos[uid + (useInstancing ? "_markers_instanced" : "_markers")]);
+    gl.bindVertexArray(glUtils._vaos[uid + (glUtils._useInstancing ? "_markers_instanced" : "_markers")]);
 
     // Set per-markerset uniforms
     gl.uniform1i(gl.getUniformLocation(program, "u_transformIndex"),
@@ -1983,33 +2010,37 @@ glUtils._drawMarkersByUID = function(gl, viewportTransform, markerScaleAdjusted,
     if (glUtils._usePiechartFromMarker[uid]) {
         // 1st pass: draw alpha for whole marker shapes
         gl.uniform1i(gl.getUniformLocation(program, "u_alphaPass"), true);
-        if (useInstancing) {
-            gl.drawArraysInstanced(gl.TRIANGLE_STRIP, 0, 4, numPoints);
-        } else {
-            gl.drawElements(gl.POINTS, numPoints, gl.UNSIGNED_INT, 0);
+        for (let offset = 0; offset < numPoints; offset += chunkSize) {
+            const count = (offset + chunkSize >= numPoints) ? numPoints - offset : chunkSize;
+            if (glUtils._useInstancing) {
+                glUtils._updateBindingOffsetsForCurrentVAO(gl, uid, offset, numPoints);
+                gl.drawArraysInstanced(gl.TRIANGLE_STRIP, 0, 4, count);
+            } else {
+                gl.drawArrays(gl.POINTS, offset, count);
+            }
         }
         // 2nd pass: draw colors for individual piechart sectors
         gl.uniform1i(gl.getUniformLocation(program, "u_alphaPass"), false);
         gl.colorMask(true, true, true, false);
-        if (useInstancing) {
-            gl.drawArraysInstanced(gl.TRIANGLE_STRIP, 0, 4, numPoints);
+        // (Reminder of the drawing is the same as for non-piechart markers, so
+        // here we can just re-use the code that follows)
+    }
+    for (let offset = 0; offset < numPoints; offset += chunkSize) {
+        const count = (offset + chunkSize >= numPoints) ? numPoints - offset : chunkSize;
+        if (glUtils._useInstancing) {
+            glUtils._updateBindingOffsetsForCurrentVAO(gl, uid, offset, numPoints);
+            gl.drawArraysInstanced(gl.TRIANGLE_STRIP, 0, 4, count);
         } else {
-            gl.drawElements(gl.POINTS, numPoints, gl.UNSIGNED_INT, 0);
-        }
-        gl.colorMask(true, true, true, true);
-    } else {
-        if (useInstancing) {
-            // Note: drawElementsInstanced is for some reason much slower than
-            // drawArraysInstanced. So since sorting currently does not work
-            // with instancing, we can use faster non-indexed drawing here.
-            gl.drawArraysInstanced(gl.TRIANGLE_STRIP, 0, 4, numPoints);
-        } else {
-            gl.drawElements(gl.POINTS, numPoints, gl.UNSIGNED_INT, 0);
+            gl.drawArrays(gl.POINTS, offset, count);
         }
     }
 
     // Restore render pipeline state
+    if (glUtils._useInstancing) {
+        glUtils._updateBindingOffsetsForCurrentVAO(gl, uid, 0, numPoints);
+    }
     gl.bindVertexArray(null);
+    gl.colorMask(true, true, true, true);
     gl.blendFunc(gl.ONE, gl.ONE);
     gl.disable(gl.BLEND);
     gl.useProgram(null);
@@ -2030,7 +2061,8 @@ glUtils._drawEdgesByUID = function(gl, viewportTransform, markerScaleAdjusted, u
     gl.uniformMatrix2fv(gl.getUniformLocation(program, "u_viewportTransform"), false, viewportTransform);
     gl.uniform2fv(gl.getUniformLocation(program, "u_canvasSize"), [gl.canvas.width, gl.canvas.height]);
     gl.uniform1f(gl.getUniformLocation(program, "u_markerScale"), markerScaleAdjusted);
-    gl.uniform1f(gl.getUniformLocation(program, "u_maxPointSize"), glUtils._useInstancing ? 2048 : 256);
+    gl.uniform1f(gl.getUniformLocation(program, "u_maxPointSize"),
+        glUtils._useInstancing ? 2048 : glUtils._caps[gl.ALIASED_POINT_SIZE_RANGE][1]);
     gl.uniform1f(gl.getUniformLocation(program, "u_edgeThicknessRatio"), glUtils._edgeThicknessRatio);
     gl.uniformBlockBinding(program, gl.getUniformBlockIndex(program, "TransformUniforms"), 0);
     gl.bindBufferBase(gl.UNIFORM_BUFFER, 0, glUtils._buffers["transformUBO"]);
@@ -2047,6 +2079,9 @@ glUtils._drawEdgesByUID = function(gl, viewportTransform, markerScaleAdjusted, u
     gl.bindTexture(gl.TEXTURE_2D, glUtils._textures[uid + "_colorLUT"]);
     gl.uniform1i(gl.getUniformLocation(program, "u_colorLUT"), 0);
 
+    // (TODO: this draw call should also be split into smaller chunks (as in
+    // glUtils._drawMarkersByUID and glUtils._drawPickingPass) to avoid problem
+    // with rendering larger datasets on some Android phones)
     gl.drawArraysInstanced(gl.TRIANGLE_STRIP, 0, 4, numEdges);
 
     // Restore render pipeline state
@@ -2084,6 +2119,7 @@ glUtils._drawRegionsColorPass = function(gl, viewportTransform) {
         gl.uniform4fv(gl.getUniformLocation(program, "u_imageBounds"), imageBounds);
         gl.uniform1i(gl.getUniformLocation(program, "u_numScanlines"), numScanlines);
         gl.uniform1f(gl.getUniformLocation(program, "u_regionOpacity"), glUtils._regionOpacity);
+        gl.uniform1f(gl.getUniformLocation(program, "u_regionStrokeWidth"), glUtils._regionStrokeWidth);
         gl.uniform1i(gl.getUniformLocation(program, "u_regionFillRule"),
             fillRuleConstants[glUtils._regionFillRule]);
         gl.uniform1i(gl.getUniformLocation(program, "u_regionUsePivotSplit"), glUtils._regionUsePivotSplit);
@@ -2120,6 +2156,11 @@ glUtils._drawRegionsColorPass = function(gl, viewportTransform) {
 
 
 glUtils._drawPickingPass = function(gl, viewportTransform, markerScaleAdjusted) {
+    // Chunk size used to split a single large draw call into smaller chunks. On
+    // some Android phones, drawing larger datasets can result in WebGL context
+    // loss or crashes, so this should be a workaround.
+    const chunkSize = 65536;
+
     // Set up render pipeline
     const program = glUtils._programs["picking"];
     gl.useProgram(program);
@@ -2129,7 +2170,8 @@ glUtils._drawPickingPass = function(gl, viewportTransform, markerScaleAdjusted) 
     gl.uniform2fv(gl.getUniformLocation(program, "u_canvasSize"), [gl.canvas.width, gl.canvas.height]);
     gl.uniform2fv(gl.getUniformLocation(program, "u_pickingLocation"), glUtils._pickingLocation);
     gl.uniform1f(gl.getUniformLocation(program, "u_markerScale"), markerScaleAdjusted);
-    gl.uniform1f(gl.getUniformLocation(program, "u_maxPointSize"), glUtils._useInstancing ? 2048 : 256);
+    gl.uniform1f(gl.getUniformLocation(program, "u_maxPointSize"),
+        glUtils._useInstancing ? 2048 : glUtils._caps[gl.ALIASED_POINT_SIZE_RANGE][1]);
     gl.uniformBlockBinding(program, gl.getUniformBlockIndex(program, "TransformUniforms"), 0);
     gl.bindBufferBase(gl.UNIFORM_BUFFER, 0, glUtils._buffers["transformUBO"]);
     gl.activeTexture(gl.TEXTURE2);
@@ -2166,7 +2208,10 @@ glUtils._drawPickingPass = function(gl, viewportTransform, markerScaleAdjusted) 
         gl.drawArrays(gl.POINTS, 0, 1);  // Note: this drawcall does not have to be indexed
         // 2nd pass: draw all the markers (as single pixels)
         gl.uniform1i(gl.getUniformLocation(program, "u_op"), 1);
-        gl.drawElements(gl.POINTS, numPoints, gl.UNSIGNED_INT, 0);
+        for (let offset = 0; offset < numPoints; offset += chunkSize) {
+            const count = (offset + chunkSize >= numPoints) ? numPoints - offset : chunkSize;
+            gl.drawArrays(gl.POINTS, offset, count);
+        }
 
         // Read back pixel at location (0, 0) to get the picked object
         const result = new Uint8Array(4);
@@ -2418,7 +2463,6 @@ glUtils.restoreLostContext = function(event) {
     for (let [uid, numPoints] of Object.entries(glUtils._numPoints)) {
         delete glUtils._buffers[uid + "_markers"];
         delete glUtils._buffers[uid + "_markers_secondary"];
-        delete glUtils._buffers[uid + "_markers_indices"];
         delete glUtils._vaos[uid + "_markers"];
         delete glUtils._vaos[uid + "_markers_instanced"];
         delete glUtils._buffers[uid + "_edges"];
@@ -2429,7 +2473,7 @@ glUtils.restoreLostContext = function(event) {
     }
 
     // Restore per-layer WebGL objects for drawing regions
-    for (let key in Objects.keys(glUtils._textures)) {
+    for (let key in Object.keys(glUtils._textures)) {
         // Check all named texture objects to be safe, since the regions
         // a texture was created for might have been deleted or moved to a
         // different image layer (been assigned a different collectionIndex)
@@ -2461,6 +2505,18 @@ glUtils.init = function() {
 
     if (!(gl instanceof WebGL2RenderingContext)) {
         interfaceUtils.alert("Error: TissUUmaps requires a web browser that supports WebGL 2.0");
+    }
+
+    // Get HW capabilities from WebGL context
+    glUtils._caps[gl.ALIASED_POINT_SIZE_RANGE] = gl.getParameter(gl.ALIASED_POINT_SIZE_RANGE);
+    console.assert(glUtils._caps[gl.ALIASED_POINT_SIZE_RANGE] instanceof Float32Array);
+
+    // Disable instanced marker drawing by default if the HW point size limit
+    // is large enough. Should be faster in most cases, and we can still
+    // temporarily switch to instanced drawing during viewport captures to
+    // avoid the HW point size limit.
+    if (glUtils._caps[gl.ALIASED_POINT_SIZE_RANGE][1] >= 1023) {
+        glUtils._useInstancing = false;
     }
 
     // Place marker canvas under the OSD canvas. Doing this also enables proper